requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.9
aiosignal==1.3.1
alembic==1.14.0
antlr4-python3-runtime==4.9.3
asteroid-filterbanks==0.4.0
attrs==24.2.0
audioread==3.0.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorlog==6.9.0
contourpy==1.3.1
cycler==0.12.1
decorator==5.1.1
docopt==0.6.2
einops==0.8.0
filelock==3.16.1
fonttools==4.55.1
frozenlist==1.5.0
fsspec==2024.10.0
huggingface-hub==0.26.3
HyperPyYAML==1.2.2
idna==3.10
Jinja2==3.1.4
joblib==1.4.2
julius==0.2.7
kiwisolver==1.4.7
lazy_loader==0.4
librosa==0.10.2.post1
lightning==2.4.0
lightning-utilities==0.11.9
llvmlite==0.43.0
Mako==1.3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.3
mdurl==0.1.2
more-itertools==10.5.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
networkx==3.4.2
numba==0.60.0
numpy==2.0.2
omegaconf==2.3.0
openai-whisper @ git+https://github.com/openai/whisper.git@90db0de1896c23cbfaf0c58bc2d30665f709f170
optuna==4.1.0
packaging==24.2
pandas==2.2.3
pillow==11.0.0
platformdirs==4.3.6
pooch==1.8.2
primePy==1.3
propcache==0.2.1
protobuf==5.29.0
pyannote.audio==3.3.2
pyannote.core==5.0.0
pyannote.database==5.1.0
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pycparser==2.22
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytorch-lightning==2.4.0
pytorch-metric-learning==2.7.0
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.12
scikit-learn==1.5.2
scipy==1.14.1
semver==3.0.2
sentencepiece==0.2.0
shellingham==1.5.4
six==1.16.0
sortedcontainers==2.4.0
soundfile==0.12.1
soxr==0.5.0.post1
speechbrain==1.0.2
SQLAlchemy==2.0.36
sympy==1.13.1
tabulate==0.9.0
tensorboardX==2.6.2.2
threadpoolctl==3.5.0
tiktoken==0.8.0
torch==2.5.1
torch-audiomentations==0.11.1
torch_pitch_shift==1.2.5
torchaudio==2.5.1
torchmetrics==1.6.0
tqdm==4.67.1
typer==0.15.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
yarl==1.18.3
